refactor reporting batch routes
