+++ src/app/router_registry.py
@@ -17,10 +17,10 @@ from app.routers.intake import router as intake_router
 from app.routers.platform import router as platform_router
 from app.routers.portfolio import router as portfolio_router
 from app.routers.proposals import router as proposals_router
-from app.routers.reporting import batches_router as reporting_batches_router
 from app.routers.reporting import router as reporting_router
-from app.routers.reporting import schedules_router as reporting_schedules_router
+from app.routers.reporting_batches import batches_router as reporting_batches_router
 from app.routers.reporting_jobs import jobs_router as reporting_jobs_router
+from app.routers.reporting_schedules import schedules_router as reporting_schedules_router
 from app.routers.reporting_snapshots import router as reporting_snapshots_router
 from app.routers.source_products import router as source_products_router
 from app.routers.workbench import router as workbench_router
